chore: Split `TitleBarAutomationPeer` into partial files
